Use ESM to use top level await and remove catches with exit
