+++ integration-tests/esbuild/build.esm.mjs
@@ -1,0 +1,7 @@
+#!/usr/bin/env node
+
+import esbuild from 'esbuild'
+
+import commonConfig from './build.esm.common-config.js'
+
+await esbuild.build(commonConfig)
